added support for new google datastore. No more ndb
