+++ pydal/dialects/__init__.py
@@ -1,9 +1,8 @@
 from .._compat import with_metaclass, iteritems
-from .._gae import gae
 from .._load import OrderedDict
 from ..helpers._internals import Dispatcher
 from ..objects import Expression
-
+from ..ndb import have_ndb
 
 dialects = Dispatcher("dialect")
 
@@ -111,6 +110,6 @@ from .oracle import OracleDialect
 from .sap import SAPDBDialect
 from .teradata import TeradataDialect
 from .couchdb import CouchDBDialect
-
-if gae is not None:
-    from .google import GoogleDatastoreDialect
+from .google import GoogleDatastoreDialect
+if have_ndb:
+    from .ndb import GoogleNDBDialect
