+++ pydal/validators.py
@@ -800,10 +800,9 @@ class IS_NOT_IN_DB(Validator):
         id = record_id or self.record_id
         if isinstance(id, dict):
             id = table(**id)
-        if not id is None:
-            query &= table._id != id
         subset = self.dbset(query, ignore_common_filters=self.ignore_common_filters)
-        if subset.select(limitby=(0, 1)):
+        rows = [row for row in subset.select(table.id, limitby=(0, 1)) if row.id != id]
+        if rows:
             raise ValidationError(self.translator(self.error_message))
         return value
